+++ protocols/v2/roles-logic-sv2/src/utils.rs
@@ -5,15 +5,16 @@
 //! management, mutex management, difficulty target calculations, merkle root calculations, and
 //! more.
 
+mod coinbase_output;
+
 use binary_sv2::U256;
 use bitcoin::Block;
 use job_declaration_sv2::{DeclareMiningJob, PushSolution};
 use primitive_types::U256 as U256Primitive;
 use std::{
     cmp::max,
-    convert::{TryFrom, TryInto},
+    convert::TryInto,
     ops::Div,
-    str::FromStr,
     sync::{Mutex as Mutex_, MutexGuard, PoisonError},
 };
 use stratum_common::{
@@ -23,13 +24,14 @@ use stratum_common::{
         consensus,
         hash_types::{BlockHash, TxMerkleNode},
         hashes::{sha256d::Hash as DHash, Hash},
-        secp256k1::{All, Secp256k1},
-        CompactTarget, PublicKey, ScriptBuf, ScriptHash, Transaction, WScriptHash, XOnlyPublicKey,
+        CompactTarget, Transaction,
     },
 };
 use tracing::error;
 
 use crate::errors::Error;
+
+pub use self::coinbase_output::CoinbaseOutput;
 
 /// Generator of unique IDs for channels and groups.
 ///
@@ -219,92 +221,6 @@ fn reduce_path<T: AsRef<[u8]>>(coinbase_id: [u8; 32], path: &[T]) -> [u8; 32] {
         root = *hash.as_ref();
     }
     root
-}
-
-/// Coinbase output transaction.
-///
-/// Typically used for parsing coinbase outputs defined in SRI role configuration files.
-#[derive(Debug, Clone)]
-pub struct CoinbaseOutput {
-    /// Specifies type of the script used in the output.
-    ///
-    /// Supported values include:
-    /// - `"P2PK"`: Pay-to-Public-Key
-    /// - `"P2PKH"`: Pay-to-Public-Key-Hash
-    /// - `"P2SH"`: Pay-to-Script-Hash
-    /// - `"P2WPKH"`: Pay-to-Witness-Public-Key-Hash
-    /// - `"P2WSH"`: Pay-to-Witness-Script-Hash
-    /// - `"P2TR"`: Pay-to-Taproot
-    pub output_script_type: String,
-
-    /// Value associated with the script, typically a public key or script hash.
-    ///
-    /// This field's interpretation depends on the `output_script_type`:
-    /// - For `"P2PK"`: The raw public key.
-    /// - For `"P2PKH"`: A public key hash.
-    /// - For `"P2WPKH"`: A witness public key hash.
-    /// - For `"P2SH"`: A script hash.
-    /// - For `"P2WSH"`: A witness script hash.
-    /// - For `"P2TR"`: An x-only public key.
-    pub output_script_value: String,
-}
-
-impl TryFrom<CoinbaseOutput> for ScriptBuf {
-    type Error = Error;
-
-    fn try_from(value: CoinbaseOutput) -> Result<Self, Self::Error> {
-        match value.output_script_type.as_str() {
-            "TEST" => {
-                let pub_key_hash = PublicKey::from_str(&value.output_script_value)
-                    .map_err(|_| Error::InvalidOutputScript)?
-                    .pubkey_hash();
-                Ok(ScriptBuf::new_p2pkh(&pub_key_hash))
-            }
-            "P2PK" => {
-                let pub_key = PublicKey::from_str(&value.output_script_value)
-                    .map_err(|_| Error::InvalidOutputScript)?;
-                Ok(ScriptBuf::new_p2pk(&pub_key))
-            }
-            "P2PKH" => {
-                let pub_key_hash = PublicKey::from_str(&value.output_script_value)
-                    .map_err(|_| Error::InvalidOutputScript)?
-                    .pubkey_hash();
-                Ok(ScriptBuf::new_p2pkh(&pub_key_hash))
-            }
-            "P2WPKH" => {
-                let w_pub_key_hash = PublicKey::from_str(&value.output_script_value)
-                    .map_err(|_| Error::InvalidOutputScript)?
-                    .wpubkey_hash()
-                    .unwrap();
-                Ok(ScriptBuf::new_p2wpkh(&w_pub_key_hash))
-            }
-            "P2SH" => {
-                let script_hashed = ScriptHash::from_str(value.output_script_value.as_ref())
-                    .map_err(|_| Error::InvalidOutputScript)?;
-                Ok(ScriptBuf::new_p2sh(&script_hashed))
-            }
-            "P2WSH" => {
-                let w_script_hashed = WScriptHash::from_str(value.output_script_value.as_ref())
-                    .map_err(|_| Error::InvalidOutputScript)?;
-                Ok(ScriptBuf::new_p2wsh(&w_script_hashed))
-            }
-            "P2TR" => {
-                // From the bip
-                //
-                // Conceptually, every Taproot output corresponds to a combination of
-                // a single public key condition (the internal key),
-                // and zero or more general conditions encoded in scripts organized in a tree.
-                let pub_key = XOnlyPublicKey::from_str(&value.output_script_value)
-                    .map_err(|_| Error::InvalidOutputScript)?;
-                Ok(ScriptBuf::new_p2tr::<All>(
-                    &Secp256k1::<All>::new(),
-                    pub_key,
-                    None,
-                ))
-            }
-            _ => Err(Error::UnknownOutputScriptType),
-        }
-    }
 }
 
 /// A list of potential errors during conversion between hashrate and target
